Add share improvements
