+++ app/src/main/java/com/travelbackintime/buybitcoin/home_coming/share/ShareHelper.kt
@@ -1,20 +1,4 @@
-/*
- * Copyright 2018 Andrey Tolpeev
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.travelbackintime.buybitcoin.home_coming.router
+package com.travelbackintime.buybitcoin.home_coming.share
 
 import android.net.Uri
 import android.support.v4.app.ShareCompat
@@ -25,31 +9,28 @@ import com.facebook.share.model.ShareLinkContent
 import com.facebook.share.widget.ShareDialog
 import com.travelbackintime.buybitcoin.home_coming.view.HomeComingFragment
 import com.travelbackintime.buybitcoin.time_travel.entity.TimeTravelResult
-import com.travelbackintime.buybitcoin.time_travel.view.TimeTravelFragment
 import com.travelbackintime.buybitcoin.utils.FormatterUtils
 import com.twitter.sdk.android.tweetcomposer.TweetComposer
 import javax.inject.Inject
 
-class HomeComingRouterImpl @Inject constructor(fragment: HomeComingFragment, private val formatterUtils: FormatterUtils) : HomeComingRouter {
+interface ShareHelper {
+
+    fun shareWithFriends(result: TimeTravelResult)
+
+    fun shareToTwitter(result: TimeTravelResult)
+
+    fun shareToFaceBook()
+}
+
+class ShareHelperImpl @Inject constructor(fragment: HomeComingFragment, private val formatterUtils: FormatterUtils) : ShareHelper {
 
     private val activity = fragment.activity as AppCompatActivity
-
-    override fun openTimeTravelFragment() {
-        val fragmentManager = activity.supportFragmentManager
-        if (fragmentManager.backStackEntryCount > 0) {
-            fragmentManager.popBackStackImmediate()
-        }
-        fragmentManager
-                .beginTransaction()
-                .replace(R.id.container, TimeTravelFragment.create())
-                .commit()
-    }
 
     override fun shareWithFriends(result: TimeTravelResult) {
         val textToShare = createShareText(result)
         ShareCompat.IntentBuilder.from(activity)
                 .setHtmlText(textToShare)
-                .setType("infoText/html")
+                .setType("text/plain")
                 .startChooser()
     }
